apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "SamvadAI - Gen-AI Complaint Management Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "django-ninja>=1.3.0",
    "django-cors-headers>=4.3.1",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
]
